Refactor: import 구조 개선
