+++ server/api/src/main.ts
@@ -3,8 +3,7 @@ import { NestFactory } from '@nestjs/core';
 import * as cookieParser from 'cookie-parser';
 import 'dotenv/config';
 import { AppModule } from './app.module';
-import { CorsConfig } from './config/cors.config';
-import { SwaggerConfig } from './config/swagger.config';
+import { CorsConfig, SwaggerConfig } from './config';
 import { HttpExceptionFilter } from './filter/http-exception.filter';
 import { TransformInterceptor } from './filter/transform.interceptor';
 import * as helmet from 'helmet';
